+++ lua/src/Lua/Ersatz/Auxiliary.hs
@@ -1,5 +1,5 @@
 {-|
-Module      : Foreign.Lua.Ersatz.Auxiliary
+Module      : Lua.Ersatz.Auxiliary
 Copyright   : © 2007–2012 Gracjan Polak;
               © 2012–2016 Ömer Sinan Ağacan;
               © 2017-2021 Albert Krewinkel
@@ -10,16 +10,16 @@ Portability : non-portable (depends on GHC)
 
 Raw bindings to ersatz functions of the auxiliary library.
 -}
-module Foreign.Lua.Ersatz.Auxiliary
+module Lua.Ersatz.Auxiliary
   ( -- * Auxiliary Library
     hsluaL_newstate
   , hsluaL_tolstring
   ) where
 
 import Foreign.C (CChar, CInt (CInt), CSize)
-import Foreign.Lua.Types (StackIndex)
+import Lua.Types (StackIndex)
 import Foreign.Ptr (Ptr)
-import qualified Foreign.Lua.Types as Lua
+import qualified Lua.Types as Lua
 
 -- | Creates a new Lua state and set extra registry values for error
 -- bookkeeping.